Add EPG management commands
Description: This commit adds new commands to manage the EPG feature of JioTV Go. The `epg generate` command generates the EPG by downloading the latest EPG from JioTV and saving it to epg.xml.gz. The `epg delete` command deletes the existing EPG file if it exists, disabling the EPG on the server. These commands provide convenient ways to enable and disable the EPG feature.