Add reset-login command to reset login credentials
This commit adds a new command "reset-login" to the main.go file. The command allows users to reset their login credentials by deleting the existing login file if it exists. This will force users to login again. The implementation of the command is done in the resetLogin.go file, which is also included in this commit.